I18n for filter buttons and clicks adjusts
