+++ app/views/companies/_filters.html.erb
@@ -11,26 +11,8 @@
                   <% params[:filter].values.each_with_index do |filter, index| %>
                     <div class="item-filter">
                       <ul>
-                        <li class="primary-nivel-filter"><button></button></li>
-                        <li class="primary-nivel-filter-secondary"><button><%= filter[:type] %> </button></li>
-                        <li class="second-nivel-filter">
-                          <div class="btn-group bootstrap-select">
-                            <button type="button" class="btn dropdown-toggle btn-default" data-toggle="dropdown" role="button" title="<%= filter[:operator] %>">
-                              <span class="filter-option pull-left"><%= filter[:operator] %></span>&nbsp;
-                              <span class="bs-caret"><span class="caret"></span></span>
-                            </button>
-                            <div class="dropdown-menu open" role="combobox">
-                              <ul class="dropdown-menu inner" role="listbox" aria-expanded="false">
-                                <li data-original-index="0" class="selected">
-                                  <a tabindex="0" class="" data-tokens="null" role="option" aria-disabled="false" aria-selected="true">
-                                    <span class="text"><%= filter[:operator] %></span>
-                                    <span class="glyphicon glyphicon-ok check-mark"></span>
-                                  </a>
-                                </li>
-                              </ul>
-                            </div>
-                          </div>
-                        </li>
+                        <li class="primary-nivel-filter-secondary"><button><%= t("filters.#{filter[:type]}") %> </button></li>
+                        <li class="second-nivel-filter"><button><%= t("filters.operators.#{filter[:operator]}") %></button></li>
                         <li class="last-nivel-filter">
                           <input type="hidden" name="filter[<%= index %>][type]" value="<%= filter[:type] %>" />
                           <input type="hidden" name="filter[<%= index %>][operator]" value="<%= filter[:operator] %>" />
@@ -85,6 +67,7 @@
       </div>
       <div class="modal-body">
         <div class="filter-geral">
+          <div class="filter-value-placeholder" hidden><%= t('filters.value') %></div>
           <div class="level-filter main-category">
             <ul>
               <li class="active">
@@ -104,29 +87,32 @@
             <ul>
               <li>
                 <a href="" data-autocomplete-category="Name"
-                           data-autocomplete-source="/companies/names">
+                           data-autocomplete-source="/companies/names",
+                           id="name">
                   <%= t('filters.name') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
               <li>
                 <a href="" data-autocomplete-category="Location"
-                           data-autocomplete-source="/companies/locations">
+                           data-autocomplete-source="/companies/locations"
+                           id="location">
                   <%= t('filters.location') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
               <li>
-                <a href="">
+                <a href="" id="description">
                   <%= t('filters.description') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
               <li>
-                <a href="">
+                <a href="" id="employees_count">
                   <%= t('filters.employees') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
               <li>
                 <a href="" data-autocomplete-category="Status"
-                           data-autocomplete-data=<%= Company::STATUSES.to_json %>>
+                           data-autocomplete-data=<%= Company::STATUSES.to_json %>
+                           id="status">
                   <%= t('filters.status') %> <i class="glyphicon glyphicon-chevron-right"></i>
                 </a>
               </li>
@@ -134,10 +120,10 @@
           </div>
           <div class="level-filter last-level-filter">
             <ul>
-              <li><a href="" data-type="equal"><%= t('filters.operators.equal') %></a></li>
-              <li><a href=""><%= t('filters.operators.contains') %></a></li>
-              <li><a href=""><%= t('filters.operators.greater') %></a></li>
-              <li><a href=""><%= t('filters.operators.less') %></a></li>
+              <li><a href="" data-type="equal" id="equal"><%= t('filters.operators.equal') %></a></li>
+              <li><a href="" id="contains"><%= t('filters.operators.contains') %></a></li>
+              <li><a href="" id="greater_than"><%= t('filters.operators.greater') %></a></li>
+              <li><a href="" id="less_than"><%= t('filters.operators.less') %></a></li>
             </ul>
           </div>
         </div>
